// fe_next/src/types/enum.ts
export enum ScoringMode {
  ACCURACY = "accuracy",
  SPEED = "speed",
}
export enum RoleBase {
  ADMIN = "ROLE_ADMIN",
  TEACHER = "ROLE_TEACHER",
  STUDENT = "ROLE_STUDENT",
}

export enum RoleBaseCode {
  ADMIN = -1,
  TEACHER = 0,
  STUDENT = 1,
}

export enum RoleBaseId {
  ADMIN = 1,
  STUDENT = 2,
  TEACHER = 3,
  BASE = -1,
}
// permissions.enum.ts
export enum EPermission {
  // Student
  STUDENT_MANAGEMENT = "P_STUDENT_MANAGEMENT",
  STUDENT_ADD = "P_STUDENT_ADD",
  STUDENT_DELETE = "P_STUDENT_DELETE",
  STUDENT_EDIT = "P_STUDENT_EDIT",

  // Test
  TEST_MANAGEMENT = "P_TEST_MANAGEMENT",
  TEST_CREATE = "P_TEST_CREATE",
  TEST_UPDATE = "P_TEST_UPDATE",
  TEST_DELETE = "P_TEST_DELETE",
  TEST_DETAIL = "P_TEST_DETAIL",

  // Test Set
  TEST_SET_MANAGEMENT = "P_TEST_SET_MANAGEMENT",
  TEST_SET_GENERATE = "P_TEST_SET_GENERATE",
  TEST_SET_CREATE = "P_TEST_SET_CREATE",
  TEST_SET_DETAIL = "P_TEST_SET_DETAIL",
  TEST_SET_UPDATE = "P_TEST_SET_UPDATE",
  TEST_SET_DELETE = "P_TEST_SET_DELETE",

  // Subject
  SUBJECT_MANAGEMENT = "P_SUBJECT_MANAGEMENT",
  SUBJECT_CREATE = "P_SUBJECT_CREATE",
  SUBJECT_UPDATE = "P_SUBJECT_UPDATE",

  // Chapter
  CHAPTER_MANAGEMENT = "P_CHAPTER_MANAGEMENT",
  CHAPTER_CREATE = "P_CHAPTER_CREATE",
  CHAPTER_UPDATE = "P_CHAPTER_UPDATE",

  // Course
  COURSE_MANAGEMENT = "P_COURSE_MANAGEMENT",
  COURSE_CREATE = "P_COURSE_CREATE",
  COURSE_UPDATE = "P_COURSE_UPDATE",

  // Question
  QUESTION_MANAGEMENT = "P_QUESTION_MANAGEMENT",
  QUESTION_CREATE = "P_QUESTION_CREATE",
  QUESTION_UPDATE = "P_QUESTION_UPDATE",
  QUESTION_DELETE = "P_QUESTION_DELETE",
  QUESTION_DETAIL = "P_QUESTION_DETAIL",

  // Exam Class
  EXAM_CLASS_MANAGEMENT = "P_EXAM_CLASS_MANAGEMENT",
  EXAM_CLASS_VIEW = "P_EXAM_CLASS_VIEW",
  EXAM_CLASS_CREATE = "P_EXAM_CLASS_CREATE",
  EXAM_CLASS_EDIT = "P_EXAM_CLASS_EDIT",
  EXAM_CLASS_DELETE = "P_EXAM_CLASS_DELETE",

  // Scoring
  SCORING_MANAGEMENT = "P_SCORING_MANAGEMENT",
  SCORING = "P_SCORING",
  SCORING_UPLOAD = "P_SCORING_UPLOAD",
  SCORING_SAVE_RESULT = "P_SCORING_SAVE_RESULT",

  // User
  USER_MANAGEMENT = "P_USER_MANAGEMENT",
  USER_CREATE = "P_USER_CREATE",
  USER_UPDATE = "P_USER_UPDATE",
  USER_DELETE = "P_USER_DELETE",

  // Teacher
  TEACHER_MANAGEMENT = "P_TEACHER_MANAGEMENT",
  TEACHER_ADD = "P_TEACHER_ADD",
  TEACHER_DELETE = "P_TEACHER_DELETE",
  TEACHER_EDIT = "P_TEACHER_EDIT",
}
export enum ERole {
  ADMIN = "ROLE_ADMIN",
  TEACHER = "ROLE_TEACHER",
  STUDENT = "ROLE_STUDENT",
  ADMIN_DEPARTMENT = "ROLE_ADMIN_DEPARTMENT",
  ADMIN_SYSTEM = "ROLE_ADMIN_SYSTEM",
}